legal_rag_chatbot/requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
requests>=2.25.0
torch>=1.9.0 